requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
matplotlib
langchain
openai
streamlit
google-generativeai
plotly